requirements.txt
# Core dependencies
requests>=2.31.0
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=2.0.0
chromadb>=0.4.0

# Optional dependencies for better performance
openpyxl>=3.1.0  # For Excel files
python-magic>=0.4.27  # For file type detection
